+++ .eslintrc.js
@@ -1,0 +1,5 @@
+// https://docs.expo.dev/guides/using-eslint/
+module.exports = {
+  extends: "expo",
+  ignorePatterns: ["/dist/*"],
+};
